moved library files into a python package
