+++ sigma_signature/WindowsEventLogsHelper.py
