+++ sigma_signature/rules/sysmon_abusing_azure_browser_sso.yml
