+++ sigma_signature/rules/sysmon_ads_executable.yml
