+++ sigma_signature/rules/sysmon_alternate_powershell_hosts_pipe.yml
