+++ sigma_signature/rules/sysmon_apt_turla_namedpipes.yml
