+++ sigma_signature/rules/sysmon_cactustorch.yml
