+++ sigma_signature/rules/sysmon_cobaltstrike_process_injection.yml
