+++ sigma_signature/rules/sysmon_createremotethread_loadlibrary.yml
