+++ sigma_signature/rules/sysmon_cred_dump_tools_named_pipes.yml
