+++ sigma_signature/rules/sysmon_mal_namedpipes.yml
