+++ sigma_signature/rules/sysmon_password_dumper_lsass.yml
