+++ sigma_signature/rules/sysmon_raw_disk_access_using_illegitimate_tools.yml
