+++ sigma_signature/rules/sysmon_susp_powershell_rundll32.yml
