+++ sigma_signature/rules/sysmon_suspicious_remote_thread.yml
