+++ sigma_signature/rules/sysmon_wmi_event_subscription.yml
