+++ sigma_signature/rules/sysmon_wmi_susp_scripting.yml
