+++ sigma_signature/signatures.py
